finaly fixed the logical memory leak, wich was caused by behaviors editing themselfs.
The bug actualy never threw an exception, but in my mind it did. foreach loops can't edit themselfs while running

The solution, the itorator patern. Which comes with a handy iterator.remove() method.

So when a behavior now is now more valid, it gets removed.

This code is not yet complete, but I changed so much shit I thought a commit was in order